+++ tests/Fake/CircularA.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Fake;
+
+final class CircularA
+{
+    public function __construct(
+        public CircularB $b,
+    ) {
+    }
+}