+++ tests/Fake/LoggerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Fake;
+
+interface LoggerInterface
+{
+    public function log(string $message): void;
+}